refactor: Increase file modularization. Issue #5
