+++ search-service/src/postgres/visitor/mod.rs
@@ -1,6 +1,9 @@
 use crate::traits::Visitor;
+
 use crate::query_representation::intermediary::Command;
+
 use crate::query_representation::r#final::command_to_query;
+
 use anyhow::Error;
 
 mod tests;
